requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
python-decouple==3.1
requests==2.18.4
slacker==0.9.65
urllib3==1.22
websockets==4.0.1
